It does work for 3x3 :3

+++ matrix.h
@@ -3,7 +3,7 @@
 
 
 typedef struct Matrix {
-    float* data;
+    double* data;
     size_t rows;
     size_t cols;
 } Matrix;
@@ -13,11 +13,11 @@ Matrix* mCreate(size_t rows, size_t cols);
 Matrix* mIdentity(size_t size);
 void mFree(Matrix* matrix);
 Matrix* mCopy(Matrix* matrix);
-void mFill(Matrix* mat, float val);
-void mGenerate(Matrix* mat, float (*generator)(Matrix* mat, int i, int j));
+void mFill(Matrix* mat, double val);
+void mGenerate(Matrix* mat, double (*generator)(Matrix* mat, int i, int j));
 Matrix* mLoadFromFile(const char* filename);
 
-float* mRow(Matrix* mat, size_t index);
+double* mRow(Matrix* mat, size_t index);
 
 void mPrint(Matrix* mat);
 
@@ -26,23 +26,23 @@ int mIsClose(Matrix* m1, Matrix* m2);
 void mTranspose(Matrix** mat);
 Matrix* mCreateTranspose(Matrix* mat);
 
-void mScale(Matrix* mat, float scalar);
+void mScale(Matrix* mat, double scalar);
 void mAdd(Matrix* a, Matrix* b);
 Matrix* mMul(Matrix* a, Matrix* b);
 void mMulFirst(Matrix** a, Matrix* b);
 
 Matrix* mCreateMinor(Matrix* mat, int index1, int index2);
-float mDet(Matrix* mat);
+double mDet(Matrix* mat);
 
 void mReduceRows(Matrix* mat);
-float mDiagonalProduct(Matrix* mat);
-float mTrace(Matrix* mat);
+double mDiagonalProduct(Matrix* mat);
+double mTrace(Matrix* mat);
 
-float mDot(Matrix* vec1, Matrix* vec2);
+double mDot(Matrix* vec1, Matrix* vec2);
 void mNormalize(Matrix* vec);
 
-float mCalcMaxEigenSymm(Matrix* mat);
-Matrix* mCalcEigVecSymm(Matrix* mat, float maxEig);
+double mCalcMaxEigenSymm(Matrix* mat);
+Matrix* mCalcEigVecSymm(Matrix* mat, double maxEig);
 Matrix* mCalcEquivalentMinorMatrix(Matrix* mat, Matrix* x);
 
 void mSVD(Matrix* mat, Matrix** u, Matrix** s, Matrix** v);
